organize pitch code into AkaoSnesTrackPitch.cpp
