+++ src/main/formats/AkaoSnes/AkaoSnesSeq.h
@@ -169,6 +169,7 @@ public:
   uint8_t tremoloFadeDepthMidiValue(int32_t depth) const;
   void updateVibratoFade();
   void updateTremoloFade();
+  void resetPitchState();
   void beginNotePitch(uint8_t note, bool validForPitchBend);
   void resetPitchBendForNewNote();
   void setPitchEnvelope(int8_t semitones, uint8_t delay, uint8_t length);
